+++ managed-ledger/src/main/java/org/apache/bookkeeper/mledger/ManagedLedgerFactory.java
@@ -18,7 +18,10 @@
  */
 package org.apache.bookkeeper.mledger;
 
+import java.util.Collections;
 import java.util.Map;
+import java.util.NavigableMap;
+import java.util.TreeMap;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Supplier;
 import org.apache.bookkeeper.common.annotation.InterfaceAudience;
@@ -27,7 +30,11 @@ import org.apache.bookkeeper.mledger.AsyncCallbacks.DeleteLedgerCallback;
 import org.apache.bookkeeper.mledger.AsyncCallbacks.ManagedLedgerInfoCallback;
 import org.apache.bookkeeper.mledger.AsyncCallbacks.OpenLedgerCallback;
 import org.apache.bookkeeper.mledger.AsyncCallbacks.OpenReadOnlyCursorCallback;
+import org.apache.bookkeeper.mledger.impl.MetaStore;
 import org.apache.bookkeeper.mledger.impl.cache.EntryCacheManager;
+import org.apache.bookkeeper.mledger.proto.MLDataFormats;
+import org.apache.pulsar.common.naming.TopicName;
+import org.apache.pulsar.common.policies.data.PersistentOfflineTopicStats;
 
 /**
  * A factory to open/create managed ledgers and delete them.
@@ -233,4 +240,21 @@ public interface ManagedLedgerFactory {
      * @return properties of this managedLedger.
      */
     CompletableFuture<Map<String, String>> getManagedLedgerPropertiesAsync(String name);
+
+    default Map<String, ManagedLedger> getManagedLedgers() {
+        return Collections.emptyMap();
+    }
+
+    default ManagedLedgerFactoryMXBean getCacheStats() {
+        return null;
+    }
+
+
+    default void estimateUnloadedTopicBacklog(PersistentOfflineTopicStats offlineTopicStats, TopicName topicName,
+                                              boolean accurate, Object ctx) throws Exception {
+    }
+
+    default ManagedLedgerFactoryConfig getConfig() {
+        return null;
+    }
 }
